удалил EntetyManager, пофиксил авторизацию пользователей, удалил не используемые именованные запросы, вынес DTO класс в отдельный пакет, исправил id с integer на Long, удалил пустой конструктор класса Pizza, снял с класса Pizza аннотацию @Component,

+++ src/main/java/com/simbirsoft/kondratyev/ruslan/pizzeria/interfacies/KitchenRepository.java
@@ -9,6 +9,6 @@ import java.util.List;
 
 @Repository
 public interface KitchenRepository extends JpaRepository<Recipe, Long> {
-    List<Recipe> findByRecipeNumber_IdEquals(Integer id);
-    void deleteRecipeByRecipeNumber_IdEquals(Integer id);
+    List<Recipe> findByRecipeNumber_IdEquals(Long id);
+    void deleteRecipeByRecipeNumber_IdEquals(Long id);
 }
